Handle downloaded file URIs correctly
This PR implements additional functionality in the Download object to prevent errors when handling encoded and decoded file URIs:
- getOutputFileUri: returns a "file://" URI;
- getOutputFileUriAsString: returns the same URI in string format;
- getOutputFile: returns a File object pointing at the downloaded item.

For example, this fixes an issue when the name of a downloaded file
contained spaces: the code to check if the file existed used the
encoded URI directly to get the file path, which gave the wrong value
(e.g. "%20" instead of spaces).

+++ app/src/common/shared/com/igalia/wolvic/downloads/DownloadsManager.java
@@ -68,12 +68,12 @@ public class DownloadsManager {
         mExecutor = new ScheduledThreadPoolExecutor(1);
     }
 
-    public  void init() {
+    public void init() {
         mContext.registerReceiver(mDownloadReceiver, new IntentFilter(DownloadManager.ACTION_DOWNLOAD_COMPLETE));
         List<Download> downloads = getDownloads();
         downloads.forEach(download -> {
-            if (mDownloadManager != null &&
-                    !new File(UrlUtils.stripProtocol(download.getOutputFileUri())).exists()) {
+            File downloadedFile = download.getOutputFile();
+            if (mDownloadManager != null && (downloadedFile == null || !downloadedFile.exists())) {
                 mDownloadManager.remove(download.getId());
             }
         });
@@ -222,9 +222,9 @@ public class DownloadsManager {
         Download download = getDownload(downloadId);
         if (download != null) {
             if (!deleteFiles) {
-                File file = new File(UrlUtils.stripProtocol(download.getOutputFileUri()));
-                if (file.exists()) {
-                    File newFile = new File(UrlUtils.stripProtocol(download.getOutputFileUri().concat(".bak")));
+                File file = download.getOutputFile();
+                if (file != null && file.exists()) {
+                    File newFile = new File(file.getAbsolutePath().concat(".bak"));
                     file.renameTo(newFile);
                     if (mDownloadManager != null) {
                         mDownloadManager.remove(downloadId);
@@ -370,12 +370,12 @@ public class DownloadsManager {
         contentValues.put(MediaStore.Downloads.DOWNLOAD_URI, download.getUri());
 
         String mimeFromExtension = null;
-        String extension = MimeTypeMap.getFileExtensionFromUrl(download.getOutputFilePath());
+        String extension = MimeTypeMap.getFileExtensionFromUrl(download.getOutputFileUriAsString());
         if (extension != null) {
             mimeFromExtension = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
         }
 
-        File downloadedFile = new File(download.getOutputFilePath());
+        File downloadedFile = download.getOutputFile();
         Uri downloadedFileUri = Uri.fromFile(downloadedFile);
         String mimeFromFile = contentResolver.getType(downloadedFileUri);
